cloud-skills-main/Day-02/consume/requirements.txt
blinker==1.6.2
boto3==1.28.45
botocore==1.31.45
click==8.1.7
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kafka-python==2.0.2
MarkupSafe==2.1.3
python-dateutil==2.8.2
s3transfer==0.6.2
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.7
python-dotenv==1.0.0